wb-run-update: fix failing on new FITs
When using wb-run-update for USB updating, it failed
on calculating rootfs SHA1.

It looks like in new FITs rootfs offset is greater than in older
because of additional kernel and DTB images, so dd utility fails
to allocate block buffer with size of rootfs offset.

Replace it with tail for now. Maybe should switch back to dd
when it will support 'iflag=skip_bytes' (probably, in jessie or
squeeze) for more performace.

Signed-off-by: Nikita webconn Maslov <wkernelteam@gmail.com>