update: factory reset support
Factory reset procedures were moved to wb-run-update to keep
install_update script in mint condition.

It's possible we will want to perform factory reset from
Web interface, or there might be some complex procedures for
that. So I think it's a good idea to keep all these
procedures in system package, which could be installed to
controller and will be compatible with all our FITs by default.

Anyway, wb-run-update script is separately shipped in initramfs
for correct USB updates.

Signed-off-by: Nikita webconn Maslov <wkernelteam@gmail.com>