When repartitioning the disk, copy the factory-fit file to partition 4.
Provide two versions of /etc/fstab: one for the new 4-partition layout and one for the old 6-partition layout. Duplicate both in the /install script (in the FIT) and in wb-configs.

Implement a mechanism so that these two fstab files live in one place (ideally in wb-configs) and are pulled from there during repartitioning.

Enhance the backup and restoration of SSH keys during repartitioning by improving the recover_certificates() function in /install. Since the keys are generated on a secure laptop, they must be “restored” from the secure p1 partition during the repartitioning process.